fix MSVC warning
git-svn-id: http://svn.osgeo.org/qgis/trunk@8469 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/core/raster/qgsrasterlayer.h
@@ -532,7 +532,7 @@ public:
     /** \brief Wrapper for GDALComputeRasterMinMax with the estimate option */
     void computeMinimumMaximumEstimates(int theBand, double* theMinMax)
     {
-      if(0 < theBand && theBand <= getBandCount())
+      if(0 < theBand && theBand <= (int) getBandCount())
       {
         GDALRasterBandH myGdalBand = GDALGetRasterBand(mGdalDataset,theBand);
         GDALComputeRasterMinMax( myGdalBand, 1, theMinMax );
